+++ Interface/AddOns/NDui/Modules/Chat/Urlcopy.lua
@@ -6,10 +6,9 @@ local strfind, strmatch, strsub, gsub = string.find, string.match, string.sub, s
 local strsplit, strlen = string.split, string.len
 
 local IsModifierKeyDown, IsAltKeyDown, IsControlKeyDown, IsModifiedClick = IsModifierKeyDown, IsAltKeyDown, IsControlKeyDown, IsModifiedClick
-local GuildInvite, BNInviteFriend = GuildInvite, BNInviteFriend
+local BNInviteFriend = BNInviteFriend
 local CanCooperateWithGameAccount = CanCooperateWithGameAccount
 local C_BattleNet_GetAccountInfoByID = C_BattleNet.GetAccountInfoByID
-local InviteToGroup = C_PartyInfo.InviteUnit
 
 local foundurl = false
 
@@ -64,10 +63,10 @@ function module:HyperlinkShowHook(link, _, button)
 		if type == "player" then
 			local unit = strmatch(value, "([^:]+)")
 			if IsAltKeyDown() then
-				InviteToGroup(unit)
+				C_PartyInfo.InviteUnit(unit)
 				hide = true
 			elseif IsControlKeyDown() then
-				GuildInvite(unit)
+				C_GuildInfo.Invite(unit)
 				hide = true
 			end
 		elseif type == "BNplayer" then
@@ -84,7 +83,7 @@ function module:HyperlinkShowHook(link, _, button)
 				elseif IsControlKeyDown() then
 					local charName = gameAccountInfo.characterName
 					local realmName = gameAccountInfo.realmName
-					GuildInvite(charName.."-"..realmName)
+					C_GuildInfo.Invite(charName.."-"..realmName)
 					hide = true
 				end
 			end