+++ Interface/AddOns/NDui/Modules/Skins/Blizzard/FrameXML/Bags.lua
@@ -138,7 +138,6 @@ tinsert(C.defaultThemes, function()
 	end
 
 	if C.db["Bags"]["Enable"] then return end
-	if not C.db["Skins"]["DefaultBags"] then return end
 
 	for i = 1, 13 do
 		local frameName = "ContainerFrame"..i
@@ -188,6 +187,7 @@ tinsert(C.defaultThemes, function()
 	BankSlotsFrame:DisableDrawLayer("BORDER")
 	BankFrameMoneyFrameBorder:Hide()
 	BankSlotsFrame.NineSlice:SetAlpha(0)
+	BankSlotsFrame.EdgeShadows:Hide()
 
 	-- "item slots" and "bag slots" text
 	select(9, BankSlotsFrame:GetRegions()):SetDrawLayer("OVERLAY")
@@ -222,6 +222,7 @@ tinsert(C.defaultThemes, function()
 	ReagentBankFrame:DisableDrawLayer("BORDER")
 	ReagentBankFrame:DisableDrawLayer("ARTWORK")
 	ReagentBankFrame.NineSlice:SetAlpha(0)
+	ReagentBankFrame.EdgeShadows:Hide()
 
 	B.Reskin(ReagentBankFrame.DespositButton)
 	B.Reskin(ReagentBankFrameUnlockInfoPurchaseButton)