+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # NDui
 
-## [8.0.1](https://github.com/siweia/NDui/tree/8.0.1) (2024-8-9)
+## [8.1.0](https://github.com/siweia/NDui/tree/8.1.0) (2024-8-17)
 
-- Bug fixes.
+- Fix item hyperlink in communities.
+- Correct hide default stancebar if disabled.
+- Fix freeslots count in warband bank.
+- Update chargeStar for Rogue.
+- Fix tradetabs for fishing.
+- Update default skins, option to switch off quest tracker.
+- Update oUF core.
+- Add classcolors for AI unit.
+- Option to block rare notification.
+- New look for new feature tag.
+- Remove deprecated code for 11.0.0.
+- Fix garrison switch button on minimap.
+- Fix npc title on friendly nameplate.
+- Update map reveal data for TWW.
+- Remove party watcher.